Support abs2 via `Future`
To smoothly transition between the old `abs2` return value and the new one
(which differs by a factor of 3 for RGB), this introduces an internal
module `ColorVectorSpace.Future` permitting users to transition
immediately to the new behavior.

`Base.abs2` will return the old value with a depwarn; users can switch
immediately to `ColorVectorSpace.Future.abs2` to avoid it.